Split net into client and server
