+++ tests/std/net.test.luau
@@ -1,7 +1,18 @@
 local net = require("@lute/net")
+local client = require("@lute/net/client")
+local server = require("@lute/net/server")
 local test = require("@std/test")
 
 test.suite("NetTestSuite", function(suite)
+	suite:case("exposesClientAndServerSubmodules", function(assert)
+		assert.eq("function", type(net.request))
+		assert.eq("function", type(net.client.request))
+		assert.eq("function", type(client.request))
+		assert.eq("function", type(net.serve))
+		assert.eq("function", type(net.server.serve))
+		assert.eq("function", type(server.serve))
+	end)
+
 	suite:case("serveLocallyAndRequest", function(assert)
 		local server = net.serve({
 			handler = function(_req: net.ReceivedRequest): net.ServerResponse
